refactor: :technologist: change Header component to Smart and Dumb component concept
